ml-services/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
scikit-learn
python-multipart
aiofiles
